pokemon-pipeline/pyproject.toml
[tool.poetry]
name = "pokemon-pipeline"
version = "0.1.0"
description = ""
authors = ["data-gorilla <dreshald@gmail.com>"]
readme = "README.md"
packages = [{include = "pokemon_pipeline", from = "src"}]

[tool.poetry.dependencies]
python = "^3.8"
pyspark = "^3.3.2"
delta-spark = "^2.2.0"
pytest = "^7.2.2"
requests = "^2.28.2"
ipython = "^8.11.0"
chispa = "^0.9.2"
pandas = "^1.5.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = "src"
addopts = [ 
    "--import-mode=importlib",
]
